+++ src/components/qAndA/AnswersList.jsx
@@ -1,69 +1,13 @@
-const { API_URL, API_KEY } = process.env;
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Box } from '@mui/material';
 import getDate from './util/getDate';
 import Answer from './Answer';
 import SeeMore from './SeeMore';
-import axios from 'axios';
 
-export default function AnswersList({ questionID }) {
-
-  const [ answers, setAnswers ] = useState([]);
-  const [ length, setLength ] = useState(2);
-  const [ isExpanded, setExpanded ] = useState(false);
-
-  const fetchAnswers = (page = 1, count = 20) => {
-
-    axios
-      .get(`${ API_URL }/qa/questions/${ questionID }/answers`, {
-        headers: { Authorization: API_KEY },
-        params: {
-          page: page,
-          count: count
-        }
-      })
-      .then((res) => {
-        const { results } = res.data;
-
-        results
-          // sort answers by helpfulness
-          .sort((a, b) => {
-            // ensure Seller's name always at top
-            if (a.answerer_name === 'Seller' && b.answerer_name !== 'Seller') {
-              return -1;
-
-            } else if (a.answerer_name !== 'Seller' && b.answerer_name === 'Seller') {
-              return 1;
-
-            } else {
-              return b.helpfulness - a.helpfulness;
-            }
-          })
-          // (bonus): for incoming answers w/ no 'reported' prop
-          .forEach((answer) => {
-            answer.reported ||= false;
-          });
-
-        setAnswers(results);
-      })
-      .catch((err) => {
-        console.error(`Error fetching answers: ${ err }`);
-      });
-  };
-
-  useEffect(() => {
-    fetchAnswers();
-  }, [/* product, questionID */]);
-  // does not seem to resolve multiple API calls —> backlog
-
-  /*
-  // check if answers fetched correctly
-  answers.forEach((a, i) => {
-    console.log(`answer ${ i }: ${ JSON.stringify(a) }`);
-  });
-  */
+export default function AnswersList({ answers, length, setLength, isExpanded, setExpanded }) {
 
   return (
+    // max height === 50% of viewport
     <Box sx={{ maxHeight: '50vh', overflowY: 'auto' }}>
       { answers
         .slice(0, length)
