+++ src/components/qAndA/Search.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { TextField } from '@mui/material';
 
 export default function Search({ search, setSearch }) {
